+++ R/nice_lm_slopes.R
@@ -10,7 +10,9 @@
 #' @param moderator The moderating variable.
 #' @param b.label What to rename the default "b" column (e.g.,
 #' to capital B if using standardized data for it to be converted
-#' to the Greek beta symbol in the `nice_table` function).
+#' to the Greek beta symbol in the `nice_table` function). Now
+#' attempts to automatically detect whether the variables were
+#' standardized, and if so, sets `b.label = "B"` automatically.
 #' @param mod.id Logical. Whether to display the model number,
 #' when there is more than one model.
 #' @param ci.alternative Alternative for the confidence interval
@@ -120,11 +122,16 @@ nice_lm_slopes <- function(model,
   table.stats <- dplyr::rename(table.stats,
                                `Predictor (+/-1 SD)` = .data$Predictor)
 
-  if (!missing(b.label)) {
-    names(table.stats)[names(
-      table.stats
-    ) == "b"] <- b.label
+  if (all(
+    lapply(models.list, function(submodel) {
+      lapply(submodel$model, function(x) {
+        c("scaled:center", "scaled:scale") %in% names(attributes(x))
+      })
+    }) %>% unlist())) {
+    b.label <- "B"
   }
+  names(table.stats)[names(table.stats) == "b"] <- b.label
+
   if (length(models.list) > 1 & mod.id == TRUE) {
     model.number <- rep(seq_along(models.list), each = 3)
     table.stats <- stats::setNames(cbind(model.number, table.stats),